requirements.txt
-r requirements-dev.txt
gunicorn
psycopg2==2.8.6
django-bootstrap-form
requests